moved np_commerce to dedicated folder
