+++ src/core/libraries/np/np_commerce/np_commerce.cpp
